+++ inst/NEWS.md
@@ -2,10 +2,15 @@ __MEDseq: Mixtures of Exponential-Distance Models with Covariates__
 ===================================================================
 
 ### Bug Fixes & Miscellaneous Edits
-* Sped-up `"kmedoids"` initialisation via `pam` for _unweighted_ sequences  
+* Sped-up `init.z="kmedoids"` initialisation via `pam` for _unweighted_ sequences,  
   by using the _highest available_ value for the `pamonce` option,  
   based on the version number of the loaded `cluster` package.
+* `init.z` gains the options `"kmodes"` & `"kmodes2"`, though only for _unweighted_ sequences:  
+  both require the newly _suggested_ `klaR (>= 0.6-13)` package.
+* `MEDseq_control` gains the arg. `random=TRUE`, governing modal sequence position tie-breaking:  
+  old behaviour (always choosing the first candidate state) recoverable via `random=FALSE`.
 * Significant speed-ups to computation of central sequences for all `opti` settings (esp. `"mode"`).
+* Fixed `seriate` options `"clusters"`/`"both"` in `plot.MEDseq` for models with no noise component.
 * `dbs` gains the optional/experimental arg. `clusters` - no change to default.
 * Fixes to `print` & `summary` methods for `MEDgating` objects if `equalPro=TRUE`.
 * Minor fixes to `print` methods for `MEDtheta` & `MEDlambda` objects also.